+++ src/network/projet_quizz/backend/src/quizz/services/index.ts
@@ -1,5 +1,7 @@
+export { ImportCollectionHandler } from './handlers/import-collection.handler';
+export { ImportLlmHandler } from './handlers/import-llm.handler';
+export { QuizzStructureService } from './core/quizz-structure.service';
+export { QuizzWriteService } from './core/quizz-write.service';
 export { QuizzImportService } from './quizz-import.service';
 export { QuizzReadService } from './quizz-read.service';
-export { QuizzStructureService } from './quizz-structure.service';
 export { QuizzService } from './quizz.service';
-export { QuizzWriteService } from './quizz-write.service';
